Refactor students app
