+++ src/apps/requests/request_models/requestModels.py
@@ -2,7 +2,7 @@ from django.db import models
 
 from apps.shared.models import SharedModelHistorical
 from apps.shared.models.choices import RequestType
-from apps.students.students_models import Student
+from apps.students.models import Student
 from apps.teachers.teacher_models.teacherModels import Teacher
 from apps.subjects.subject_models.subjectModels import Subject
