+++ src/apps/shared/models/choices/week_days.py
@@ -1,4 +1,3 @@
-# Week Days
 class WeekDays:
     MONDAY = "LUNES"
     TUESDAY = "MARTES"
