+++ src/apps/students/api/serializers/__init__.py
@@ -1,0 +1,8 @@
+from .students_serializers import (
+    StudentViewSerializer,
+    StudentCreationSerializer,
+)
+from .relatives_serializers import (
+    RelativeViewSerializer,
+    RelativeCreationSerializer,
+)
