+++ src/apps/students/api/views/__init__.py
@@ -1,0 +1,1 @@
+from .students_view_set import StudentViewSet, RelativeViewSet
